+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Display only link (instead dropdown) if only one action
 - Fix data if no filter set
 - deliverDownload typo
+- Move DefaultBrowserFormat and DefaultSettingsStorage to Renderer
 
 ## [0.20.5]
 - check typed array interfaces